pull all the format declarations out to their own module
